+++ dmr/controller.py
@@ -10,6 +10,7 @@ from typing import (
 )
 
 from django.http import HttpRequest, HttpResponse, HttpResponseBase
+from django.urls import URLPattern
 from django.utils.functional import classproperty
 from django.views import View
 from django.views.decorators.csrf import csrf_exempt
@@ -608,10 +609,20 @@ class Controller(Blueprint[_SerializerT_co], View):  # noqa: WPS214
         )
 
     @classmethod
-    def get_path_item(cls, path: str, context: OpenAPIContext) -> PathItem:
+    def get_path_item(
+        cls,
+        path: str,
+        pattern: URLPattern,
+        context: OpenAPIContext,
+    ) -> PathItem:
         """Generate OpenAPI spec for path items."""
         operations: dict[str, Any] = {
-            method.lower(): endpoint.get_schema(path, cls.serializer, context)
+            method.lower(): endpoint.get_schema(
+                path,
+                pattern,
+                cls.serializer,
+                context,
+            )
             for method, endpoint in cls.api_endpoints.items()
         }
         return PathItem(