+++ docs/examples/components/path_msgspec.py
@@ -1,0 +1,50 @@
+from typing import Annotated
+
+import msgspec
+from django.urls import include
+from typing_extensions import TypedDict
+
+from dmr import Controller, Path
+from dmr.openapi import build_schema
+from dmr.openapi.views import OpenAPIJsonView, SwaggerView
+from dmr.plugins.msgspec import MsgspecSerializer
+from dmr.routing import Router, path
+
+
+class _PathModel(TypedDict):
+    user_id: Annotated[str, msgspec.Meta(min_length=4, max_length=4)]
+    post_id: Annotated[int, msgspec.Meta(gt=0)]
+
+
+class PostController(
+    Controller[MsgspecSerializer],
+    Path[_PathModel],
+):
+    def get(self) -> _PathModel:
+        return self.parsed_path
+
+
+router = Router(
+    'api/',
+    [
+        # We define a regular Django path:
+        path(
+            'user/<str:user_id>/post/<int:post_id>/',
+            PostController.as_view(),
+            name='user',
+        ),
+    ],
+)
+schema = build_schema(router)
+
+urlpatterns = [
+    # Register our router in the final url patterns:
+    path(router.prefix, include((router.urls, 'test_app'), namespace='api')),
+    # Add swagger:
+    path('docs/openapi.json/', OpenAPIJsonView.as_view(schema), name='openapi'),
+    path('docs/swagger/', SwaggerView.as_view(schema), name='swagger'),
+]
+
+# run: {"controller": "PostController", "method": "get", "url": "/api/user/abcd/post/1/", "use_urlpatterns": true}  # noqa: ERA001, E501
+# run: {"controller": "PostController", "method": "get", "url": "/api/user/abcd/post/0/", "use_urlpatterns": true, "curl_args": ["-D", "-"], "fail-with-body": false}  # noqa: ERA001, E501
+# openapi: {"openapi_url": "/docs/openapi.json/", "use_urlpatterns": true}  # noqa: ERA001, E501